Overhauled item slot organization
Item slots are now pointers to data. The data they point to will now be
a centralized location to write data to (ie. treasure to put there,
collect mode, & potentially future things like multiworld player).

The main advantage to this is that it's become much easier to specify
data (ie. collect mode) based on a specific item slot, instead of
creating data for a specific room which applies to all treasures in that
room. (I may still need to do that in many cases for convenience's
sake, but at least the whole system doesn't depend on that structure.)