HOOK: Disable use of switch-hook on flowers
Surprisingly, Season's "breakable tile data" already works pretty well
with the switch hook. The only problem (that I'm aware of) is that it
allows switch hooking spring flowers. This is problematic because their
graphics are located in the lower half of the graphics data, meaning
they can't be turned into sprites, causing them to look corrupted.