CANE: Increased priority of cane / mystery sound
Without this change, when you swing the cane while another block exists,
all you can hear is the "poof", not the sound of the new block being
created. This is noticeably different from Ages. By increasing the
"priority" of the cane's sound, this difference is fixed.

This may subtly affect sounds in Seasons if there are other situations
where the priority of this particular sound matters.