HARP: Added Harp of Ages to Seasons
Songs have not been ported over, but that's not difficult to do. They
make a placeholder sound.

Harp songs which are displayed on the inventory will appear on top of
the submenu that appears when opening the seed
satchel/shooter/slingshot. In Ages, this is avoided by drawing some
blank sprites that will hide harp song. But, this can't be done in
Seasons without some surprisingly intrusive changes.