RANDO-CROSSITEMS: Moved D5 magnet ball to prevent softlock
If you had the cane of somaria and not the magnet gloves, but you went
ahead and spawned the magnet ball anyway, then it would become
impossible to pass through that area (despite being able to open the
door with the cane of somaria). To prevent the softlock, the magnet ball
is moved 1 tile to the left.