HARP: Cover the harp song with sprites when opening seed submenus
As in Ages, sprites are now used to cover up the harp song to prevent it
from appearing on top of seed selection submenus.

This doesn't work perfectly - the song sprite still appears on top of
the menu as it's opening, because color 0 is used (unlike in Ages) which
always appears below sprites. But, it's good enough.