Stop using local labels prefixed with "_" (except in bank 0)
It was causing more confusion than it was helping, because deciding what
should be local and what shouldn't be often felt pretty arbitraty - and
it's also a bit ugly to have "_" prefixing everything. Now that we're
using namespaces in most places, using local labels doesn't help that
much anyway. But we're still using them in bank 0 since any junk in
there will be visible in all contexts.