HACK-BASE: Change build directory name
It came to my attention that the room layout files were not being
rebuild after switching from an editable "master" branch build. Perhaps
I could find some workaround for that, but at the end of the day it is
probably safer to just have them use a different build directory.