HACK-BASE: Stopped using layout groups
The "layout group" is a per-tileset value which determines which "group"
of rooms to load from (roomXXYY.bin, where XX is the layout group). This
is very annoying:

- Easy to get confused with "wActiveGroup" which is what we normally
mean by "group".
- In LynnaLab, multiple rooms in different "groups" could share the same
"layout group", meaning they'd have the same room layout. Confusing.
- Because layout groups are specified per tileset, each tileset can only
be used within one particular "world". There are some duplicate tilesets
for precisely this reason.

This is solved by simply hardcoding each "group" to a specific "layout
group", which almost always works. But, some events which change the
layouts of certain areas had to be rewritten to modify the layout group
specifically instead of the whole tileset.

Natzu Prairie in Seasons was the trickiest thing to fix with this new
system; the room layout files for those had to be shuffled around. See
tools/hack-base/moveNatzuRooms.sh for details.