Changed format of warp data
Also split into two separate files for dest data and source data.

Once again, using the new system for "continue bits".

As it turns out, the reason the Ages warp data looked slightly different
was that they needed to disable the last defined warp always being the
default one if no data was found. This was necessary in order for
the fallback dungeon stair warp detection code to work.